feat: share runtime context
