+++ packages/playground-ui/src/domains/agents/index.tsx
@@ -4,3 +4,4 @@ export * from './components/agent-traces';
 export * from './context';
 export * from './components/agent-settings';
 export * from './components/agent-table/agent-table';
+export * from './components/runtime-context';
